post oldest entries first
we are assuming the feed is in reverse chronological order here, which
may be incorrect. we'd need to sort the feed by date to fix that,
something we're not doing right now anyways.

this also makes the code simpler because we can simply write the cache
file after every post, which makes it possible to resume safely after
a crash (assuming that cPickle doesn't corrupt its data file, which
isn't the case, but that was a problem before anyways)