automatically configure oauth access tokens
we still rely on provided consumer keys and secrets, but at least now
access tokens are dynamically generated. we also add a --init option
to dynamically regenerate those.

the --init flag will look like it's configuring tokens multiple times
if you use the `-c ~/.config/feed2tweet.ini` option to work around
again, of course.

to test this, I used:

    feed2tweet.py -n --init

i haven't tested with a new post, but it should behave correctly as
well.

this is to help automating the creation of the config file. at some
point, feed2tweet would get an official consumer key that can then be
configured as a default in the app to remove the need for the user to
manually configure an app in twitter.